Small fixes
Made it that content scripts wont be injected if the url has "dontInjectScripts". Used for when we scrape an list.
Added an text animation on buttons in options page.
Added an check for unknown usernames
Made an main function in StatsScraper that creates the window and starts scraping and after its done we can easily close it on StatsScraper.