Done     Features         Custom entry title             - Allows you to modify the title of an entry and by pressing enter you can toggle between the orginal and modified title. Why i made this is becauase sometimes you can't find the manga using animeplanet's title.
        Show filtering
            - Shows filter options for screens that are smaller than 768px wide.

        Cleaner alt titles
            - Splits alt titles from commas into blocks (Previously this code was in 'Extra manga data' feature)

    Added
        Made modifications to options page to make it look/run better on mobile.
        Added an new key that dictates if an content script works on mobile or desktop.

    BUGS
        Set more precise selector for notes because before if there was an ad it would mess up the layout of notes.
        Script tracking was not working correctly so i remade the way i am tracking scripts.

It seems animeplanet does not add the tooltips to dom when your user agents indicate you are on phone, which makes a lot of features not work on mobile. I have been trying to find a method that would fix this issue but hitting a roadblock after roadblock with it. I could always just fetch the animeplanet site with a custom user agent and show that fetched animeplanet site, but I would prefer not to do that instead, I will try to see if I can find a better method.